+++ tests/e2e/test_step_types.py
@@ -22,13 +22,13 @@ from tests.conftest import SPECIFICATIONS_DIR
                 "step_4_python_pandas",
             ],
         ),
-        # Expanded pipeline with loops, parallel splits, hierarchy
+        # Expanded pipeline with loops, clones, hierarchy
         (
             "e2e/pipeline_expanded.yaml",
             [
-                "step_1_parallel_split_1_step_1_python_pandas",
-                "step_1_parallel_split_2_step_1_python_pandas",
-                "step_1_parallel_split_3_step_1_python_pandas",
+                "step_1_clone_1_step_1_python_pandas",
+                "step_1_clone_2_step_1_python_pandas",
+                "step_1_clone_3_step_1_python_pandas",
                 "step_2_python_pandas",
                 "step_3_loop_1_step_3_python_pandas",
                 "step_3_loop_2_step_3_python_pandas",
@@ -76,7 +76,7 @@ def test_step_types(
     # Check that we get directories for particular implementations
     for implementation in implementations:
         assert (test_specific_results_dir / "diagnostics" / implementation).exists()
-        # step_3 is embarrassingly parallel so the final results land in _aggregate/
+        # step_3 is auto-parallel so the final results land in _aggregate/
         aggregate_mapping = {
             "step_3_python_pandas": "step_3_aggregate",
             "step_3_loop_1_step_3_python_pandas": "step_3_loop_1_aggregate",